Resolved circular dependencies in change event
